Make web.py dummier
This change moves all logic into MergifyWorker

Queue hash and methods pickled by rq are now in queue.py

web.py now just drops messages into redis and nothing more.

This allow us to retry anything without requiring to use the Github UI
to resend events.
